fix stream
