+++ app/src/app/function/helper.rs
@@ -800,118 +800,3 @@ pub trait FunctionCallHelper: UseJobExecutor + McpNameConverter + Send + Sync {
     }
 }
 
-// /// Listen to job result by job_id and return a static-lifetime stream of JSON values.
-// ///
-// /// This is a standalone function that takes owned/Arc parameters, suitable for use
-// /// in spawned tasks that require 'static lifetime.
-// ///
-// /// # Arguments
-// /// * `function_app` - Arc to FunctionAppImpl (or any FunctionCallHelper)
-// /// * `job_id` - The job ID to listen for (owned)
-// /// * `runner_name` - Runner name for decoding output (owned)
-// /// * `using` - Optional method name for multi-method runners (owned)
-// /// * `timeout` - Optional timeout in milliseconds
-// ///
-// /// # Returns
-// /// Stream of JSON values representing the job result chunks with 'static lifetime
-// pub fn listen_job_result_as_json_stream_static<F>(
-//     function_app: std::sync::Arc<F>,
-//     job_id: proto::jobworkerp::data::JobId,
-//     runner_name: String,
-//     using: Option<String>,
-//     timeout: Option<u64>,
-// ) -> std::pin::Pin<Box<dyn futures::Stream<Item = Result<serde_json::Value>> + Send + 'static>>
-// where
-//     F: FunctionCallHelper + Send + Sync + 'static,
-// {
-//     use futures::StreamExt;
-//     use proto::jobworkerp::data::result_output_item;
-
-//     // UseJobResultApp is inherited from FunctionCallHelper -> UseJobExecutor -> UseJobResultApp
-//     Box::pin(async_stream::stream! {
-//         let result = function_app
-//             .job_result_app()
-//             .listen_result_by_job_id(&job_id, timeout, true)
-//             .await;
-
-//         match result {
-//             Ok((_job_result, Some(mut stream))) => {
-//                 // Stream available - yield decoded chunks
-//                 while let Some(item) = stream.next().await {
-//                     match item.item {
-//                         Some(result_output_item::Item::Data(data)) => {
-//                             match function_app
-//                                 .decode_job_result_output(None, Some(&runner_name), &data, using.as_deref())
-//                                 .await
-//                             {
-//                                 Ok(json) => yield Ok(json),
-//                                 Err(e) => {
-//                                     tracing::warn!("Failed to decode stream chunk: {:?}", e);
-//                                     // Fallback to raw string
-//                                     yield Ok(serde_json::Value::String(
-//                                         String::from_utf8_lossy(&data).to_string(),
-//                                     ));
-//                                 }
-//                             }
-//                         }
-//                         Some(result_output_item::Item::End(_)) |
-//                         Some(result_output_item::Item::FinalCollected(_)) => {
-//                             // End of stream
-//                             break;
-//                         }
-//                         None => {
-//                             // Skip empty items
-//                         }
-//                     }
-//                 }
-//             }
-//             Ok((job_result, None)) => {
-//                 // No stream - job already completed, use final result
-//                 tracing::debug!(
-//                     "No stream available for job {} (already completed), using final result",
-//                     job_id.value
-//                 );
-
-//                 if let Some(data) = &job_result.data {
-//                     if let Some(output) = &data.output {
-//                         if !output.items.is_empty() {
-//                             match function_app
-//                                 .decode_job_result_output(
-//                                     None,
-//                                     Some(&runner_name),
-//                                     &output.items,
-//                                     using.as_deref(),
-//                                 )
-//                                 .await
-//                             {
-//                                 Ok(json) => yield Ok(json),
-//                                 Err(e) => {
-//                                     tracing::warn!("Failed to decode final result: {:?}", e);
-//                                     // Fallback to raw string
-//                                     yield Ok(serde_json::Value::String(
-//                                         String::from_utf8_lossy(&output.items).to_string(),
-//                                     ));
-//                                 }
-//                             }
-//                         } else {
-//                             yield Err(JobWorkerError::RuntimeError(
-//                                 "Job completed but output is empty".to_string(),
-//                             ).into());
-//                         }
-//                     } else {
-//                         yield Err(JobWorkerError::RuntimeError(
-//                             "Job completed but output is None".to_string(),
-//                         ).into());
-//                     }
-//                 } else {
-//                     yield Err(JobWorkerError::RuntimeError(
-//                         "Job completed but has no result data".to_string(),
-//                     ).into());
-//                 }
-//             }
-//             Err(e) => {
-//                 yield Err(e);
-//             }
-//         }
-//     })
-// }
