+++ proto/protobuf/jobworkerp/data/workflow_event.proto
@@ -93,6 +93,19 @@ message WorkflowCancelledEvent {
   string position = 2;
 }
 
+// # Streaming Data Event
+// Event emitted for each chunk of streaming data (e.g., LLM token)
+// Used for real-time streaming to clients (e.g., AG-UI TEXT_MESSAGE_CONTENT)
+message StreamingDataEvent {
+  // # Job ID
+  // Unique identifier of the streaming job
+  int64 job_id = 1;
+
+  // # Data
+  // Raw streaming data chunk (typically UTF-8 encoded text)
+  bytes data = 2;
+}
+
 // # Workflow Event
 // Union type representing all possible events during workflow execution
 // Used for streaming workflow progress to clients (e.g., AG-UI)
@@ -129,6 +142,11 @@ message WorkflowEvent {
     // # Generic Task Completed
     // Emitted when ForTask, SwitchTask, DoTask, etc. finishes
     TaskCompletedEvent task_completed = 6;
+
+    // # Streaming Data
+    // Emitted for each chunk of streaming data (e.g., LLM token)
+    // Used for real-time streaming to AG-UI clients
+    StreamingDataEvent streaming_data = 7;
 
     // # Workflow Cancelled (Reserved)
     // Emitted when workflow is cancelled
